refactor(GiniHealthAPILibrary): Remove pinning projects
