+++ HealthAPILibrary/GiniHealthAPILibraryExample/GiniHealthAPILibraryExample/ViewController.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import GiniHealthAPILibrary
-import GiniHealthAPILibraryPinning
 
 class ViewController: UIViewController {
     override func viewDidLoad() {
