--- HealthAPILibrary/GiniHealthAPILibraryPinning/.gitignore
@@ -1,7 +1,0 @@
-.DS_Store
-/.build
-/Packages
-/*.xcodeproj
-xcuserdata/
-DerivedData/
-.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
