--- HealthAPILibrary/GiniHealthAPILibraryPinning/LICENSE
@@ -1,12 +1,0 @@
-Copyright (c) 2019-2021, Gini GmbH
-All rights reserved.
-
-The Gini Health API Library Pinning is licensed through Gini GmbH ("Gini") and may not be
-used, altered or copied in any way without explicit permission by Gini. The
-terms of usage are defined in a separate usage agreement between Gini and the
-licensee, where the licensee can gain access to a non-exclusive,
-non-transferable usage right which is restricted for the time of a contractual
-relationship between Gini and the licensee.
-
-For license related inquiries contact Gini via the email address
-technical-support@gini.net.
