--- HealthAPILibrary/GiniHealthAPILibraryPinning/README.md
@@ -1,31 +1,0 @@
-<p align="center">
-<img src="./GiniHealth_Logo.png" width="250">
-</p>
-
-# Gini Health API Library Pinning for iOS
-
-[![Platform](https://img.shields.io/badge/platform-iOS-lightgrey.svg)]()
-[![Swift version](https://img.shields.io/badge/swift-5.0-orange.svg)]()
-[![Swift package manager](https://img.shields.io/badge/Swift_Package_Manager-compatible-orange?style=flat-square)]()
-
-The Gini Health API Library Pinning provides ways to interact with the Gini Health API and therefore, adds the possiblity to scan documents and extract information from them. The library supports certificate pinning.
-
-## Documentation
-
-Further documentation with information about how install and integrate it can be found in our [website](https://developer.gini.net/gini-mobile-ios/GiniHealthAPILibrary/index.html).
-
-## Requirements
-
-- iOS 12+
-- Xcode 12+
-
-## Author
-
-Gini GmbH, hello@gini.net
-
-## License
-
-The Gini Health API Library Pinning for iOS is licensed under a Private License. See [the license](https://developer.gini.net/gini-mobile-ios/GiniHealthAPILibrary/license.html) for more info.
-
-> ⚠️  **Important:** Always make sure to ship all license notices and permissions with your application.
-
