--- HealthAPILibrary/GiniHealthAPILibraryPinning/Sources/GiniHealthAPILibraryPinning/GiniHealthAPILibraryPinningVersion.swift
@@ -1,8 +1,0 @@
-//
-//  GiniHealthAPILibraryPinningVersion.swift
-//  
-//
-//  Copyright © 2024 Gini GmbH. All rights reserved.
-//
-
-public let GiniHealthAPILibraryPinningVersion = "4.2.0"
