--- HealthAPILibrary/GiniHealthAPILibraryPinning/Tests/GiniHealthAPILibraryPinningTests/GiniHealthAPILibraryPinningTests.swift
@@ -1,7 +1,0 @@
-import XCTest
-@testable import GiniHealthAPILibraryPinning
-
-final class GiniHealthAPILibraryPinningTests: XCTestCase {
-    func testExample() throws {
-    }
-}
