--- HealthSDK/GiniHealthSDKExample/GiniHealthSDKPinningExample/de.lproj/Localizable.strings
@@ -1,7 +1,0 @@
-/* 
-  Localizable.strings
-  HealthSDKPinningExample
-
- //  Copyright © 2024 Gini GmbH. All rights reserved.
-
-*/
