--- HealthSDK/GiniHealthSDKExample/GiniHealthSDKPinningExample/en.lproj/Localizable.strings
@@ -1,8 +1,0 @@
-/* 
-  Localizable.strings
-  HealthSDKPinningExample
-
- //  Copyright © 2024 Gini GmbH. All rights reserved.
-
-*/
-"ginihealth.reviewscreen.iban.placeholder" = "IBANXXXX";
