--- HealthSDK/GiniHealthSDKPinning/.swiftpm/xcode/xcuserdata/home.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>GiniHealthSDKPinning.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
-		<key>GiniHealthSDKPinningTests.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>9</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>GiniHealthSDKPinning</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>GiniHealthSDKPinningTests</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
