--- HealthSDK/GiniHealthSDKPinning/Sources/GiniHealthSDKPinning/GiniHealthSDKPinningVersion.swift
@@ -1,8 +1,0 @@
-//
-//  GiniHealthSDKPinningVersion.swift
-//  
-//
-//  Copyright © 2024 Gini GmbH. All rights reserved.
-//
-
-public let GiniHealthSDKPinningVersion = "4.2.0"
