--- HealthSDK/GiniHealthSDKPinning/Tests/GiniHealthSDKPinningTests/HealthSDKPinningTests.swift
@@ -1,7 +1,0 @@
-import XCTest
-@testable import GiniHealthSDKPinning
-
-final class GiniHealthSDKPinningTests: XCTestCase {
-    func testExample() throws {
-    }
-}
