--- MerchantSDK/GiniMerchantSDK/Sources/GiniMerchantSDK/Core/SSLPinning/GiniSessionDelegate.swift
@@ -1,22 +1,0 @@
-//
-//  GiniSessionDelegate.swift
-//
-//
-//  Copyright © 2024 Gini GmbH. All rights reserved.
-//
-
-import Foundation
-
-class GiniSessionDelegate: NSObject, URLSessionDelegate {
-    private let pinningManager: SSLPinningManager
-    
-    internal init(pinningConfig: [String: [String]]) {
-        self.pinningManager = SSLPinningManager(pinningConfig: pinningConfig)
-    }
-    
-    func urlSession(_ session: URLSession,
-                    didReceive challenge: URLAuthenticationChallenge,
-                    completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
-        pinningManager.validate(challenge: challenge, completionHandler: completionHandler)
-    }
-}
